pdm-api-types: remote upid: add type field to RemoteUpid
In quite a few places in the code where we handle RemoteUpids, we need
to know the actual type of the remote, for instance to
  - build correct API paths from it, if the UPID is passed to a
    product-specific API endpoint
  - get fields from the actual UPID, which requires parsing the native
    UPID type - if the type is unknown here, we have to either guess
    (attempt to parse one type, and if it does not work, try the other
    type), or get the type from somewhere else (some other parameter, or
    from remotes.cfg)

These can be easily solved by storing the type of the remote in the
RemoteUpid. The serialized representation is changed in such a way that
the type is prepended to the original represenation, e.g.

  pve:remote-name!<original UPID>

This change aims to avoid breakage by being backward-compatible with the
old representation without the type field. In this case the type is
simply inferred by parsing the UPID. This adds some runtime cost, but
this is only really relevant for migrating the contents of the remote
task cache over to the new format. Once it has been migrated (which
happens automatically when rewriting the archive files, or if that does
not happen, they are simply rotated out after a while), the inefficient
code path is not really needed any more.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Reviewed-by: Shannon Sterz <s.sterz@proxmox.com>
Tested-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20251111105059.148997-6-l.wagner@proxmox.com