resources: views: support new pve network resource type
Proxmox VE now returns the new network resource type, in addition to
the sdn resource type. This patch adds support for this new format.

Since PDM makes one API call to a single remote (`/cluster/resources`)
to obtain the resources, handling of the different resource types is
already done by the Proxmox VE remote. If it is on a version that uses
the sdn resource type, then only sdn resources will get returned and
vice vers.

For that reason, the whole output can simply be deserialized and used
without any special case handling on PDM side. Both resource types get
mapped to the same PDM-internal struct. An additional legacy flag has
been added to the zone resource, in order for us to be able to
distinguish whether the resource came from an older or newer PVE
version. This is required to generate the correct URLs for the
resources in the frontend, among other things.

Distinguishing between different types of network entities, now
happens via the network_type, field of network resources. Instead of
returning a specific type like `sdn-zone`, use the same type for all
network resources and provide filtering. This will be implemented in a
successive patch.

There is no need for providing a backwards-compatible Resource struct
for deserialization purposes, since the only place where the current
sdn resources are used, use the ephemeral resource cache - which gets
deleted upon restarting the service.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251113162112.795835-4-s.hanreich@proxmox.com