ui: use new network resource type
Instead of an SDN resource, the backend now returns the network
resource. Fix all occurences of this enum in the UI code, so they use
the new variant.

In addition, network entities can now be filtered via the network-type
search category. Adjust the SDN zone panel from the Dashboard, so it
supports this new match category, when filtering for SDN zones.

Depending on the current version of the remote, the URL structure for
inspecting resources has changed. By utilizing the legacy parameter
returned by the backend, we can decide which link to generate in the
SDN zone tree.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251113162112.795835-6-s.hanreich@proxmox.com