fix: add details button to show subscriptions dialog
When clicking on the (www.proxmox.com) url on the subscriptions panel
the url will open and also the dialog window at the same time. Inorder
to fix the issue add a "Details" button on the subscriptions panel. Now
the dialog will only be shown when the "Details" button is clicked.

Signed-off-by: Shan Shaji <s.shaji@proxmox.com>
Link: https://lore.proxmox.com/20251114181242.423922-1-s.shaji@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>