remote tasks: correctly filter out foreign running PBS tasks
At the moment, only tasks started by PDM will ever show up as 'active' -
mostly to avoid having to regularly poll foreign (as in, not started by
PDM) tasks. If we included running tasks in the task cache, they could
appear as 'stuck' for up to one poll interval, even if they ran only for
a couple of seconds.

This commit fixes a mistake when checking whether the task is still
running or not. Before, we *only* included running tasks, which is the
opposite of what we want.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20251117153509.390137-2-l.wagner@proxmox.com