Revert "server: use types indead of string for migration parameters"
The problem with using the flattened PVE type directly is that the
`target-endpoint` it expects differs from what PDM wants as
`target-endpoint`. Specififically, PDM expects it to be an optional
string identifying the target node. The PVE type however has a
PropertyString<ProxmoxRemote> as `target-endpoint`, which has two
mandatory fields (host and api-token). The only way to have this working
with the PVE type is to have the UI set `target-endpoint` to something
like:
`host=192.168.1.34,apitoken=dummy`
It does not really allow for auto-select node to be encoded properly.

This reverts commit dbb8524549c335987046ebe6e742635e1357aa3d.

Signed-off-by: Hannes Laimer <h.laimer@proxmox.com>
Link: https://lore.proxmox.com/20251118122105.119918-2-h.laimer@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>