api: migrate: use arrays for storage and bridge mappings
With [1] we generate actual arrays for string-lists, so the generated
functions we use for migration take these types containing Vec's instead
of Strings. This updates the PDM migrate endpoints to accept arrays so
they can be passed to the functions.

This also updates the strigify function used by the UI to create those
mapping strings. Now, instead of a single string it produces a list of
mapping strings.

[1] bd8eca6 ("pve-api-types: schema2rust: generate arrays for types with format `-list`")

Signed-off-by: Hannes Laimer <h.laimer@proxmox.com>
Link: https://lore.proxmox.com/20251118122105.119918-3-h.laimer@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>