client: let task status command infer remote for upid
Since the remote is a separate parameter, and the `task list` command
does not actually include the remote in the UPIDs it prints, let's
automatically add the remote if the passed UPID does not contain it.

This way we can copy-paste UPIDs from the `task list` output into a
`task status` command without having to type the remote twice and
fumbling a `!` separator in manually.

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>