tree-wide: add enum fallback variants for pbs api types
pbs-api-types added fallback variants to all enums that are part of
the REST API. Fix all occurences of the enums that got new fallback
variants in PDM by handling them properly. In the backend, unknown
values are logged, while in the frontend they're simply rendered with
a special icon / font color.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251120145031.550340-7-s.hanreich@proxmox.com
 [TL: update d/control]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>