lib: api-types: add 'layout' property to ViewConfig
this is a simple string that holds the layout json. We can't currently
add it as a normal property with 'correct' api types, since we want to
use enum features that are not available with the api macro.

Multiple reasons why we can't use the correct type here (currently):
* our api macro does not support enum variants with struct types
  (so e.g. `Foo { bar: usize }`) so we can't generate an api schema for
  it.
* using something generic like a `Value` does not work since that does
  not impl our UpdaterType and we can't here because of the orphan rule.
* we could try to use a wrapper type around `Value` and implement
  the `UpdaterType` and `ApiType` ourselves, but the section config
  won't parse more complex types like an `ObjectSchema`

So until these are possible, our best bet is to simply use a string
and be careful what we accept (like parsing into the wanted type
during parsing/updating).

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251126151833.3637080-3-d.csapak@proxmox.com