ui: dashboard: prepare view for editing custom views
This adds a mechanism to edit a view, namely it adds an edit button
(pencil) in the status row. When in 'edit mode' one can:
* drag the panels around
* delete panels
* add new panels
* set the 'flex' value of panels
* add a new row at the end
* delete a whole row

There is currently no mechanism to persistently save the result, but
that's only a case of wiring the 'on_update_layout' callback to e.g. a
backend api call.

Also the editing is only active when the view is not named 'dashboard'.

The drag&drop works with desktop and touchscreens, but on touchscreens,
there is no 'drag item' shown currently.

The menu structure for adding new items could probably be improved, but
that should not be a big issue.

For handling the 'editing overlay' of the panels, there is a new
'RowElement' component that just abstracts that away to have a less
code in the RowView component.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251126151833.3637080-9-d.csapak@proxmox.com