ui: resource tree: move error message into first column
this makes the 'node' column much more predictable, so we
can set a fixed width that should be enough for most nodenames instead
of using up much space (double than the 'id' one).

To properly display the error, we can use the 'colspan' attribute when
we use `render_cell` in the first column.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251126151833.3637080-22-d.csapak@proxmox.com