api-types/ui: add ResourceTree variant for WidgetType
this is the resource tree from the search, but contained within a panel
and with search field in a toolbar, for use in the custom views.

This can come in handy, since it gives an overview of the filtered
resources.

It does not refresh together with the other panels in the view
currently, only once at the beginning and when the search term changes.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251126151833.3637080-25-d.csapak@proxmox.com