pdm-api-types: views: preparations for future glob/regex support
Change the config format slightly in a way similar to how the
'match-field' statements work for notification matchers.

The new format is, in pseudo-regex:

  (include|exclude) (exact:)?(resource-pool|...|tag)=.*

The 'exact:' part is optional. Later we can add "regex" or "glob" types
to add support for globbing or regex without any changes to the config
format.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20251126155859.413048-2-l.wagner@proxmox.com