api: firewall: add option, rules and status endpoints
This adds the following endpoints
* for all PVE remotes:
 - GET /pve/firewall/status

* for PVE remotes
 - GET pve/remotes/{remote}/firewall/options
 - PUT pve/remotes/{remote}/firewall/options
 - GET pve/remotes/{remote}/firewall/rules
 - GET pve/remotes/{remote}/firewall/status

* for PVE node
 - GET pve/remotes/{remote}/nodes/{node}/firewall/options
 - PUT pve/remotes/{remote}/nodes/{node}/firewall/options
 - GET pve/remotes/{remote}/nodes/{node}/firewall/rules
 - GET pve/remotes/{remote}/nodes/{node}/firewall/status

* for guests (both lxc and qemu)
 - GET pve/remotes/{remote}/[lxc|qemu]/{vmid}/firewall/options
 - PUT pve/remotes/{remote}/[lxc|qemu]/{vmid}/firewall/options
 - GET pve/remotes/{remote}/[lxc|qemu]/{vmid}/firewall/rules

`options` endpoints are for recieving and updating the configured
firewall options for remotes, nodes and guests. Both lxc and qemu guests
share the same type for getting and upating their options.

`rules` endpoints return the list of firewall rules that exist on the
entity. All remotes, nodes and guests return a list with items of the
same type.

`status` endpoints return the firewall status of the entity, this
includes:
 - name/id
 - optional status(enabled, count of enabled rules)
 - list of 'child-statuses', so:
    for pve status (all remotes) -> list of remote-statuses
    for remote status -> list of node-statuses
    for node status -> list of guest-statuses
    for guest status -> no list
 -(only guest) type of guest

Like this we have a way to limit the amount of requests the PDM has to
make in order to collect all the needed data. Given the rather large
amoutn of requests needed to assemble all the data this made more sense
than always loading everything and filtering on the client side.

Data to build the status response is fetch in parallel using our
ParallelFetcher. But only  for `all remotes` or `single remote`
status requests. The status for single nodes is done sequentially.

Signed-off-by: Hannes Laimer <h.laimer@proxmox.com>
Link: https://lore.proxmox.com/20251127092209.213535-7-h.laimer@proxmox.com