ui: add firewall status tree
Adds tree displaying the firewall status of remotes, nodes and guests.
Upon selecting an entity in the tree the right panle shows its
configured firewall rules, this is read-only, so rules can't be added,
deleted or modified currently. The tree contains a button that allows to
edit the firewall options of remotes, nodes or guests.

Given the rather large amount of requests it takes PDM to accumulate all
the data this contains comboboxes to select specific remotes or nodes
on a specific remote. This doesn't just filter the results but changes
how data is requested. So is is possible to limit the amount of work the
PDM has to do for each request and improve responsiveness. The text
filter is only for local, already loaded, data, so it won't trigger a new
load.

Signed-off-by: Hannes Laimer <h.laimer@proxmox.com>
Link: https://lore.proxmox.com/20251127092209.213535-9-h.laimer@proxmox.com