pdm-api-types: reuse APTUpdateInfo from proxmox_apt_api_types
... instead of defining of defining our own version locally.
Both structs are equivalent, with the exception of `change_log_url`,
which was present in the pdm-api-types version but not in the struct
definition in proxmox-apt-api-types. We never seem to use this member
(grepped PBS as well as PDM source), so it should be fine to use the
definition without it - it also still compiles fine.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20251127104447.162951-7-l.wagner@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>