ui: remote updates: don't attempt to load current status for unavailable nodes
If the cached update status indicates that a node is unavailable, don't
attempt to load the repo status and update table. Instead, show the
error message from the last poll attempt in-place where the tables would
else be. This avoids annoying error pop-ups when navigating through the
remote/node tree in the global update view.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20251127104447.162951-17-l.wagner@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>