ui: dashboard: subscriptions: refactor subscriptions show logic
Instead of creating the dialog in the click callback and passing it
around, just to clone it again in the view. Simply set a boolean flag
and create the dialog in the view method if needed.

This is more in line with what we do with most dialogs and saves us a
few headaches later (since the dialog will be rendered on every `view`
call instead of just once).

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Reviewed-by: Michael Köppl <m.koeppl@proxmox.com>
Tested-by: Michael Köppl <m.koeppl@proxmox.com>
Link: https://lore.proxmox.com/20251127140451.3131469-2-d.csapak@proxmox.com