ui: main: abort running task load on log out
loading currently running tasks wasn't aborted on logout. this is due
to the Loader not being dropped as the DatacenterManagerApp is always
rendered and it never drops its loader. this could trigger a race
condition between a load with a potentially expired ticket and a new
log in request. if the load returned after the login, a user would be
kicked out right after logging in again.

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20251127153609.440415-3-s.sterz@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>