ui: main: only render acl context when we are logged in
otherwise it will only log errors about not being able to load the
current users acls. the request to load the acl tree could also end up
in a race with a login attempt. if the context loading with an invalid
ticket and a log in request end up running in parallel, a user will be
kicked out after logging in if the context is done loading after the
login request.

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20251127153609.440415-5-s.sterz@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>