server: add jobstate module from PBS
This is a general helper module that helps with keeping track of
periodic tasks. Thus far we have not really needed it in PDM (although
it could be useful for remote update fetching), but this changes with
task log rotation which will be added next.

The module was copied as is and then adapted where needed to make it
work with PDM; some types and paths were changed. In the future this
should be broken out and added to some shared crate (could be a new
micro crate if no better place comes to mind), but given the time
constraints a copy is good enough for now. This module will probably
remain pretty unchanged in the near future, so the risk of diverging
implementations should be pretty small.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20251127153708.393210-2-l.wagner@proxmox.com