ui: pve qemu: merge VM config tabs into single one
In the same spirit as the previous commit for LXC merged the four tabs
there, also merge the two tabs for QEMU VMs for now, i.e.:

Having that many tabs crowds the view and needs much more clicking to
iterate through all of them. Besides, we do not use the same base
layout as in PVE, so we also do not have to split the config into the
same views as there.

By merging those four tabs into a single one, users get a good
overview of the VM much quicker; albeit for some users seeing all of
the config options at once might be a bit much. For that I'd rather
split the hardware section into general (cpu, memory, display, bios,
...) and a dedicated disk section and a dedicated network one, i.e. as
it's long time planned to do for PVE.

Independent of that: we might want to rethink this _if_ we enable
editing in the future due to having multiple top-bars with the edit
buttons around.
But for the time being we also do not plan to make this fully editable
here, besides maybe very central limits, and editing those could, for
example, be exposed through an in-row edit icon.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>