server: api: sdn: add ip-vrf endpoint
Calls the respective Proxmox VE endpoint to obtain status information
about the IP-VRF of an EVPN zone. Since the status is per-node, use
the existing nodes subdirectory instead of the general SDN
subdirectory, mirroring the API paths used in Proxmox VE. In order to
avoid having too many SDN-specific modules for a single API call,
introduce a submodule in the sdn module directly. If needed, this can
be easily factored out in the future.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Tested-by: Hannes Dürr <h.duerr@proxmox.com>
Reviewed-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20251128110905.104189-5-s.hanreich@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>