ui: sdn: evpn: add vnet status panel
This panel shows the status of the MAC-VRF of an EVPN vnet. It lists
the neighbor table of the vnet on a given node, which can be selected
via the dropdown in the panel. It will be used in the remote tree and
the vrf tree of the EVPN view to display detailed information about
the selected EVPN vnet.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Tested-by: Hannes Dürr <h.duerr@proxmox.com>
Reviewed-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20251128110905.104189-8-s.hanreich@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>