sdn: evpn: add detail panel to the evpn panel
Extend the EVPN panel, so it can display detailed information about
selected elements in the trees. When selecting a specific zone / vnet,
the detail panel will show the IP / MAC-VRF of that zone / vnet. This
requires some refactoring of the existing EVPN panel to allow
displaying a second panel next to it. It is now structured like the
remote view, which also offers a panel that shows an overview of the
remote and details of the selected node.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Tested-by: Hannes Dürr <h.duerr@proxmox.com>
Reviewed-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20251128110905.104189-9-s.hanreich@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>