remote tasks: poll foreign, non-tracked active tasks to avoid them getting stuck
If tasks which are currently active but not tracked (not started by PDM)
are added to the task cache, then under some special circumstances, they
can get stuck in the 'active' state. This happened mostly due to a bug
that was already fixed [1]. As a safeguard and to fix existing stuck
tasks, we now poll active tasks with a long interval (10min) and check
if they are finished.

This polling is done as part of the regular poll loop and does not
result in additional API calls if there are no active, non-tracked
tasks.

[1] fixed in 6247ff3c7

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20251128140522.311838-2-l.wagner@proxmox.com