api: remote tasks: add fine-grained permission check on the remote level
Adds a check for the remote task APIs that requires a user to have
Resource.Audit on /resource/{remote}.

Could make sense to a.) verify permissions on the node-level (we do so
in PVE) and b.) allow a user to see their own tasks. But we can do that
later as well, more important for now is getting rid of the bare
Permission::Anybody.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20251128151550.369432-1-l.wagner@proxmox.com