server: add system report implementation
This code was taken from PBS and then adapted for PDM. While it could
make sense to refactor and then share some of the helper functions with
PBS, the benefit is rather small (the helpers are relatively trivial)
and now is not the best time for it. The risk and potential consequences
of diverging implementations is small.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Tested-by: Stefan Hanreich <s.hanreich@proxmox.com>
Reviewed-by: Shannon Sterz <s.sterz@proxmox.com>
Tested-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20251201153229.347742-2-l.wagner@proxmox.com