resource tree: fix navigating to network resources
The resources tree generated `sdn/zones` as a slug for the zone
network resource, which shows the SDN resource tree - but the actual
corect slug would be `sdn`. Additionally, with the addition of
fabrics to the network resources, make the resource tree navigate to
the SDN tree as well, instead of just for zones.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251201175724.73576-1-s.hanreich@proxmox.com