ui: remote updates: ensure apt repo status uses correct subscription information
The AptRepositories component supported taking a base_url for the repo
api call but had the one for subscription info hard coded from the
"/nodes/localhost/subscription", which meant that we mixed the remote
repo status with the PDMs subscription status, resulting in rather
confusing displays.

That component now gained a new property to define that URL, so set
that to ensure the correct info gets used when calculating the repo
status report.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>