ui: rename PBS node slug to overview as stop-gap for search routing
This makes sure that the correct view is rendered when searching and
clicking on a PBS datastore while being on a panel not related to the
PBS remotes.

The search produces the correct path in navigate_to() when the
PdmResourceTree triggers Msg::NavigateToEntry, setting this via
the navigator(). The result is however still being the node status
view slug to be appended, which leads to datastores being displayed
incorrectly. By setting the slug to `overview`, it now works for both
cases.

This is intended as temporary stop-gap until the actual routing
behaviour is fixed.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20251202083152.71130-1-c.ebner@proxmox.com