lib/server: pve: add api call to get the cached version info from remotes
this is useful when we e.g. want to feature gate some things in the ui.
This way we don't have to get the whole update information for all
remotes, and we don't have to query all nodes of the remotes for their
versions.

Should we need it in the future, we can do the analagous api call for
pbs too.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251202111533.2068448-4-d.csapak@proxmox.com