api-types: resource: fix storage identifier
Storages in PVE are uniquely identified by their name and the node
they reside on. The current method only includes the name of the
storage, not the node. Since the ID method is used in the UI for
generating the deep URLs to the respective PVE resources, the
generated URLs are wrong.

The generated global ID is correct, because the map_pve_storage
function uses the id field returned by the PVE API directly, which
includes the node name already.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251202143809.475821-1-s.hanreich@proxmox.com